Added option vary_commutations; if yes, vary on all commutations of commutable gates to find shortest circuit.
The call to this new optimization should be in the back-end compiler, and before the back-end scheduler.
An example of such a call was added to the arch/cc_light/cc_light_eqasm_compiler.h
It only has effect with the post179 scheduler and when the scheduler_commute option is yes.